+++ src/beam/transformer/reducer.py
@@ -1,0 +1,12 @@
+from beam.processor import Processor
+from beam.utils import collate_chunks
+
+
+class Reducer(Processor):
+
+    def __init__(self, hparams, *args, dim=1, **kwargs):
+        super().__init__(hparams, *args, **kwargs)
+        self.dim = dim
+
+    def reduce(self, *xs, **kwargs):
+        return collate_chunks(*xs, dim=self.dim, **kwargs)
